+++ Sources/SwifterSwift/Foundation/LocaleExtensions.swift
@@ -26,6 +26,25 @@ public extension Locale {
         let dateString = dateFormatter.string(from: Date())
         return dateString.contains(dateFormatter.amSymbol) || dateString.contains(dateFormatter.pmSymbol)
     }
+
+}
+
+// MARK: - Functions
+public extension Locale {
+
+    /// SwifterSwift: Get the flag emoji for a given country region code.
+    /// - Parameter isoRegionCode: The IOS region code.
+    ///
+    /// Adapted from https://stackoverflow.com/a/30403199/1627511
+    static func flagEmoji(forRegionCode isoRegionCode: String) -> String? {
+        guard isoRegionCodes.contains(isoRegionCode) else { return nil }
+
+        return isoRegionCode.unicodeScalars.reduce(into: String()) {
+            guard let flagScalar = UnicodeScalar(UInt32(127397) + $1.value) else { return }
+            $0.unicodeScalars.append(flagScalar)
+        }
+    }
+
 }
 
 #endif